requirements.txt
asgiref==3.4.1
asn1crypto==1.4.0
base58==2.1.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.10.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.28.1
idna==3.3
ipfshttpclient==0.7.0
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
multiaddr==0.0.9
mysqlclient==2.0.3
netaddr==0.8.0
oauthlib==3.1.1
oscrypto==1.2.1
Pillow==8.4.0
pycparser==2.20
pyHanko==0.9.0
pyhanko-certvalidator==0.17.3
PyJWT==1.7.1
python-barcode==0.13.1
python-pkcs11==0.7.0
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
uharfbuzz==0.16.1
uritemplate==4.1.1
uritools==3.0.2
urllib3==1.26.7
varint==1.0.2
